per reviewer request, move non-scalar test classes
